Remove indirection in several structs through DSTs
Some types are always behind a box/rc, and having internally boxed dyn
trait fields is an unnecessary redirection. These types are now unsized
and have the dyn trait inlined within the struct, for those sweet
nanoseconds of extra performance.

+++ game/procgen/src/planet.rs
@@ -338,7 +338,7 @@ impl Planet {
         &self,
         chunks: impl Iterator<Item = ChunkLocation>,
         z_range: (GlobalSliceIndex, GlobalSliceIndex),
-        mut per_point: impl FnMut(u64, WorldPosition),
+        mut per_point: impl FnMut(usize, WorldPosition),
     ) {
         let inner = self.0.read().await;
         for region in chunks
