+++ game/world/src/loader/loading.rs
@@ -487,7 +487,7 @@ impl<C: WorldContext> WorldLoader<C> {
         &self,
         chunks: impl Iterator<Item = ChunkLocation>,
         z_range: (GlobalSliceIndex, GlobalSliceIndex),
-        per_point: impl FnMut(u64, WorldPosition),
+        per_point: impl FnMut(usize, WorldPosition),
     ) {
         let fut = self
             .source
